`target_type` is no more (#33)
target_type was an enum that was built from the combination of section_type
and compartments (this was a bool).

Now we have 2 enums that are passed by string in report.conf. This has several
benefits:

- inspection: we can understand more easily what we are passing
- de-correlation: we can change one enum without changing the other
- no need of special encoding/decoding: before we needed to pack and
  unpack the enums into 1
- easier to expand in case more options arise

Drawbacks:

- less performant since we need to encode/decode with strings

I think the drawback are totally ininfluent. Report parsing is not
performance critical and clarity and simplicity should rein here

* fix write